# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: VJU Club Management API
  version: 1.0.0
  description: 'API giai đoạn 1 cho quản lý tài khoản, CLB, ban, thành viên và permission.
    Xem x-implementation-status từng operation: IMPLEMENTED là đã xác nhận qua integration
    test. UUID, camelCase, thời gian ISO-8601. List dùng offset/limit; không có page/size.'
servers:
- url: /api/v1
  description: Backend cùng origin với Swagger UI
- url: http://localhost:8080/api/v1
  description: Backend local mặc định
security:
- sessionAuth: []
paths:
...
  /clubs/{clubId}/memberships:
    get:
      operationId: membership.list
      summary: List club memberships
      tags:
      - Membership
      x-permission: member.view
      x-implementation-status: IMPLEMENTED
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipListResponse'
              example:
                items:
                - id: 44444444-4444-4444-8444-444444444444
                  userId: 11111111-1111-4111-8111-111111111111
                  clubId: 22222222-2222-4222-8222-222222222222
                  status: ACTIVE
                  joinedAt: '2026-09-29T12:00:00Z'
                  leftAt: null
                total: 1
                offset: 0
                limit: 20
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
          - LEFT
          - SUSPENDED
        description: Lọc theo trạng thái. Bỏ trống để lấy cả lịch sử (các membership đã LEFT).
      - name: departmentId
        in: query
        required: false
        schema:
          type: string
          format: uuid
        description: Chỉ lấy membership đang được phân vào ban này.
      - name: search
        in: query
        required: false
        schema:
          type: string
        description: Tìm theo email, họ tên hoặc mã sinh viên.
      - name: offset
        in: query
        schema:
          type: integer
          minimum: 0
          default: 0
        description: Số bản ghi bỏ qua; bắt đầu từ 0.
      - name: limit
        in: query
        schema:
          type: integer
          minimum: 1
          default: 20
          maximum: 100
        description: Số bản ghi tối đa trong một lần lấy, mặc định 20, tối đa 100.
    post:
      operationId: membership.add
      summary: Add member
      tags:
      - Membership
      x-permission: member.add
      x-implementation-status: IMPLEMENTED
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Membership'
              example:
                id: 44444444-4444-4444-8444-444444444444
                userId: 11111111-1111-4111-8111-111111111111
                clubId: 22222222-2222-4222-8222-222222222222
                status: ACTIVE
                joinedAt: '2026-09-29T12:00:00Z'
                leftAt: null
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMembershipRequest'
            example:
              userId: 11111111-1111-4111-8111-111111111111
      description: >-
        Thêm user vào CLB. Mỗi lần tham gia là một membership mới; user đã rời (LEFT)
        có thể được thêm lại. Trả 409 MEMBERSHIP_ALREADY_EXISTS nếu user đang có membership
        chưa kết thúc, CLUB_INACTIVE hoặc USER_INACTIVE.
  /memberships/{membershipId}:
    get:
      operationId: membership.get
      summary: Get membership
      tags:
      - Membership
      x-permission: member.view_detail
      x-implementation-status: IMPLEMENTED
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Membership'
              example:
                id: 44444444-4444-4444-8444-444444444444
                userId: 11111111-1111-4111-8111-111111111111
                clubId: 22222222-2222-4222-8222-222222222222
                status: ACTIVE
                joinedAt: '2026-09-29T12:00:00Z'
                leftAt: null
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
      parameters:
      - name: membershipId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      operationId: membership.update
      summary: Update membership
      tags:
      - Membership
      x-permission: member.update
      x-implementation-status: IMPLEMENTED
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Membership'
              example:
                id: 44444444-4444-4444-8444-444444444444
                userId: 11111111-1111-4111-8111-111111111111
                clubId: 22222222-2222-4222-8222-222222222222
                status: ACTIVE
                joinedAt: '2026-09-29T12:00:00Z'
                leftAt: null
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
      parameters:
      - name: membershipId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMembershipRequest'
            example:
              status: SUSPENDED
      description: >-
        Đổi trạng thái membership. LEFT là trạng thái cuối: chuyển sang LEFT sẽ gỡ user
        khỏi mọi ban của CLB; membership đã LEFT không mở lại được (409 MEMBERSHIP_ALREADY_LEFT).
    delete:
      operationId: membership.remove
      summary: Remove member
      tags:
      - Membership
      x-permission: member.remove
      x-implementation-status: IMPLEMENTED
      responses:
        '204':
          description: Member removed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
      parameters:
      - name: membershipId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      description: Đưa thành viên ra khỏi CLB bằng trạng thái LEFT và ghi leftAt.
        Không xóa user. Thành viên ACTIVE mới được phân vào ban.
  /departments/{departmentId}/members:
    get:
      operationId: departmentMember.list
      summary: List department members
      tags:
      - Departmentmember
      x-permission: department.member.view
      x-implementation-status: IMPLEMENTED
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartmentMemberListResponse'
              example:
                items:
                - id: 66666666-6666-4666-8666-666666666666
                  departmentId: 33333333-3333-4333-8333-333333333333
                  membershipId: 44444444-4444-4444-8444-444444444444
                  userId: 11111111-1111-4111-8111-111111111111
                  joinedAt: '2026-09-29T12:00:00Z'
                total: 1
                offset: 0
                limit: 20
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
      parameters:
      - name: departmentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: offset
        in: query
        schema:
          type: integer
          minimum: 0
          default: 0
        description: Số bản ghi bỏ qua; bắt đầu từ 0.
      - name: limit
        in: query
        schema:
          type: integer
          minimum: 1
          default: 20
          maximum: 100
        description: Số bản ghi tối đa trong một lần lấy, mặc định 20, tối đa 100.
    post:
      operationId: departmentMember.add
      summary: Assign department member
      tags:
      - Departmentmember
      x-permission: department.member.add
      x-implementation-status: IMPLEMENTED
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartmentMember'
              example:
                id: 66666666-6666-4666-8666-666666666666
                departmentId: 33333333-3333-4333-8333-333333333333
                membershipId: 44444444-4444-4444-8444-444444444444
                userId: 11111111-1111-4111-8111-111111111111
                joinedAt: '2026-09-29T12:00:00Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
      parameters:
      - name: departmentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddDepartmentMemberRequest'
            example:
              membershipId: 44444444-4444-4444-8444-444444444444
  /departments/{departmentId}/members/{membershipId}:
    patch:
      operationId: departmentMember.move
      summary: Move department member
      tags:
      - Departmentmember
      x-permission: department.member.remove & department.member.add
      x-implementation-status: IMPLEMENTED
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartmentMember'
              example:
                id: 66666666-6666-4666-8666-666666666666
                departmentId: 33333333-3333-4333-8333-333333333333
                membershipId: 44444444-4444-4444-8444-444444444444
                userId: 11111111-1111-4111-8111-111111111111
                joinedAt: '2026-09-29T12:00:00Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
      parameters:
      - name: departmentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: membershipId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveDepartmentMemberRequest'
            example:
              targetDepartmentId: 33333333-3333-4333-8333-333333333333
      x-permission-rules:
        sourceDepartment: department.member.remove
        targetDepartment: department.member.add
      description: departmentId trên URL là ban nguồn. Body targetDepartmentId là
        ban đích, cùng CLB. Cần department.member.remove tại nguồn và department.member.add
        tại đích. Xóa nguồn/thêm đích trong cùng transaction; lỗi thì giữ nguồn.
    delete:
      operationId: departmentMember.remove
      summary: Remove department member
      tags:
      - Departmentmember
      x-permission: department.member.remove
      x-implementation-status: IMPLEMENTED
      responses:
        '204':
          description: Removed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
      parameters:
      - name: departmentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: membershipId
        in: path
        required: true
        schema:
          type: string
          format: uuid
...
  /users/me/applications:
    get:
      operationId: clubApplication.listMine
      summary: List my club applications
      description: Chỉ trả đơn của người gọi; hỗ trợ trạng thái và CLB, phân trang offset/limit,
        sắp xếp mới nhất trước. Không cần permission quản trị.
      tags: [ClubApplication]
      x-permission: null
      x-implementation-status: IMPLEMENTED
      parameters:
      - name: status
        in: query
        schema: {type: string, enum: [PENDING, APPROVED, REJECTED, CANCELLED]}
      - name: clubId
        in: query
        schema: {type: string, format: uuid}
      - name: sort
        in: query
        schema: {type: string, default: createdAt,desc}
        description: Sắp xếp theo createdAt với hướng asc hoặc desc, ví dụ createdAt,desc.
      - name: offset
        in: query
        schema: {type: integer, minimum: 0, default: 0}
      - name: limit
        in: query
        schema: {type: integer, minimum: 1, maximum: 100, default: 20}
      responses:
        '200':
          description: My applications
          content:
            application/json:
              schema: {$ref: '#/components/schemas/MyClubApplicationListResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
...
  /users/me/memberships:
    get:
      operationId: membership.listMine
      summary: List my club membership history
      description: Chỉ trả membership của người gọi; gồm club, trạng thái, joinedAt và leftAt.
        Không cần permission quản trị và sắp xếp theo joinedAt mới nhất trước.
      tags: [Membership]
      x-permission: null
      x-implementation-status: IMPLEMENTED
      parameters:
      - name: status
        in: query
        schema: {type: string, enum: [ACTIVE, INACTIVE, LEFT, SUSPENDED]}
      - name: offset
        in: query
        schema: {type: integer, minimum: 0, default: 0}
      - name: limit
        in: query
        schema: {type: integer, minimum: 1, maximum: 100, default: 20}
      responses:
        '200':
          description: My memberships
          content:
            application/json:
              schema: {$ref: '#/components/schemas/MyMembershipListResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /users/me/notifications:
    get:
      operationId: notification.listMine
      summary: List my in-app notifications
      description: Chỉ trả thông báo của người gọi, sắp xếp mới nhất trước.
      tags: [Notification]
      x-permission: null
      x-implementation-status: IMPLEMENTED
      parameters:
      - name: offset
        in: query
        schema: {type: integer, minimum: 0, default: 0}
      - name: limit
        in: query
        schema: {type: integer, minimum: 1, maximum: 100, default: 20}
      responses:
        '200':
          description: My notifications
          content:
            application/json:
              schema: {$ref: '#/components/schemas/NotificationListResponse'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: CLUB_SESSION
      description: 'Cookie phiên do POST /auth/login đặt. Request thay đổi dữ liệu cần thêm header X-XSRF-TOKEN.'
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: 'Chưa đăng nhập hoặc phiên đã hết hạn/bị huỷ (code UNAUTHORIZED); đăng nhập lại.'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Permission denied
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: Trùng dữ liệu hoặc trạng thái hiện tại không cho phép thao tác.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Không tìm thấy resource hoặc grant.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    MyClubApplicationListResponse:
      type: object
      required: [items, total, offset, limit]
      properties:
        items:
          type: array
          items:
            type: object
            required: [id, clubId, clubCode, clubName, status, createdAt]
            properties:
              id: {type: string, format: uuid}
              clubId: {type: string, format: uuid}
              clubCode: {type: string}
              clubName: {type: string}
              status: {type: string, enum: [PENDING, APPROVED, REJECTED, CANCELLED]}
              createdAt: {type: string, format: date-time}
              reviewedAt: {type: [string, 'null'], format: date-time}
              cancelledAt: {type: [string, 'null'], format: date-time}
        total: {type: integer, format: int64}
        offset: {type: integer}
        limit: {type: integer}
    MyMembershipListResponse:
      type: object
      required: [items, total, offset, limit]
      properties:
        items:
          type: array
          items: {$ref: '#/components/schemas/MyMembership'}
        total: {type: integer, format: int64}
        offset: {type: integer}
        limit: {type: integer}
    MyMembership:
      type: object
      required: [id, clubId, clubCode, clubName, status, joinedAt, createdAt, updatedAt, departments]
      properties:
        id: {type: string, format: uuid}
        clubId: {type: string, format: uuid}
        clubCode: {type: string}
        clubName: {type: string}
        status: {type: string, enum: [ACTIVE, INACTIVE, LEFT, SUSPENDED]}
        joinedAt: {type: string, format: date-time}
        leftAt: {type: [string, 'null'], format: date-time}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
        departments:
          type: array
          items:
            type: object
            required: [id, name, joinedAt]
            properties:
              id: {type: string, format: uuid}
              name: {type: string}
              joinedAt: {type: string, format: date-time}
    Notification:
      type: object
      required: [id, title, message, read, createdAt]
      properties:
        id: {type: string, format: uuid}
        title: {type: string, maxLength: 160}
        message: {type: string, maxLength: 2000}
        read: {type: boolean}
        referenceId: {type: [string, 'null'], format: uuid}
        createdAt: {type: string, format: date-time}
    NotificationListResponse:
      type: object
      required: [items, total, offset, limit]
      properties:
        items:
          type: array
          items: {$ref: '#/components/schemas/Notification'}
        total: {type: integer, format: int64}
        offset: {type: integer}
        limit: {type: integer}
...
    Membership:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        userId:
          type: string
          format: uuid
        clubId:
          type: string
          format: uuid
        status:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
          - LEFT
          - SUSPENDED
        joinedAt:
          type: string
          format: date-time
          readOnly: true
        leftAt:
          type:
          - string
          - 'null'
          format: date-time
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
      required:
      - id
      - userId
      - clubId
      - status
      - joinedAt
      - leftAt
      - createdAt
      - updatedAt
      examples:
      - id: 44444444-4444-4444-8444-444444444444
        userId: 11111111-1111-4111-8111-111111111111
        clubId: 22222222-2222-4222-8222-222222222222
        status: ACTIVE
        joinedAt: '2026-09-29T12:00:00Z'
        leftAt: null
        createdAt: '2026-09-29T12:00:00Z'
        updatedAt: '2026-09-29T12:00:00Z'
    MoveDepartmentMemberRequest:
      type: object
      required:
      - targetDepartmentId
      properties:
        targetDepartmentId:
          type: string
          format: uuid
      examples:
      - targetDepartmentId: 33333333-3333-4333-8333-333333333333
    DepartmentMember:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        departmentId:
          type: string
          format: uuid
        membershipId:
          type: string
          format: uuid
        joinedAt:
          type: string
          format: date-time
          readOnly: true
        userId:
          type: string
          format: uuid
      required:
      - id
      - departmentId
      - membershipId
      - joinedAt
      - userId
      examples:
      - id: 66666666-6666-4666-8666-666666666666
        departmentId: 33333333-3333-4333-8333-333333333333
        membershipId: 44444444-4444-4444-8444-444444444444
        userId: 11111111-1111-4111-8111-111111111111
        joinedAt: '2026-09-29T12:00:00Z'
...
    Problem:
      type: object
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
          description: URI request, có thể vắng với lỗi do security filter trả về.
        code:
          type: string
      required:
      - type
      - title
      - status
      - detail
      - code
      examples:
      - type: about:blank
        title: PERMISSION_DENIED
        status: 403
        detail: Permission denied
        instance: /api/v1/clubs/22222222-2222-4222-8222-222222222222
        code: PERMISSION_DENIED
...
    MembershipListResponse:
      type: object
      required:
      - items
      - total
      - offset
      - limit
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Membership'
        total:
          type: integer
          format: int64
        offset:
          type: integer
        limit:
          type: integer
      examples:
      - items:
        - id: 44444444-4444-4444-8444-444444444444
          userId: 11111111-1111-4111-8111-111111111111
          clubId: 22222222-2222-4222-8222-222222222222
          status: ACTIVE
          joinedAt: '2026-09-29T12:00:00Z'
          leftAt: null
        total: 1
        offset: 0
        limit: 20
    DepartmentMemberListResponse:
      type: object
      required:
      - items
      - total
      - offset
      - limit
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DepartmentMember'
        total:
          type: integer
          format: int64
        offset:
          type: integer
        limit:
          type: integer
      examples:
      - items:
        - id: 66666666-6666-4666-8666-666666666666
          departmentId: 33333333-3333-4333-8333-333333333333
          membershipId: 44444444-4444-4444-8444-444444444444
          userId: 11111111-1111-4111-8111-111111111111
          joinedAt: '2026-09-29T12:00:00Z'
        total: 1
        offset: 0
        limit: 20
...
    CreateMembershipRequest:
      type: object
      required:
      - userId
      properties:
        userId:
          type: string
          format: uuid
      examples:
      - userId: 11111111-1111-4111-8111-111111111111
    UpdateMembershipRequest:
      type: object
      required:
      - status
      properties:
        status:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
          - LEFT
          - SUSPENDED
      examples:
      - status: SUSPENDED
    AddDepartmentMemberRequest:
      type: object
      required:
      - membershipId
      properties:
        membershipId:
          type: string
          format: uuid
      examples:
      - membershipId: 44444444-4444-4444-8444-444444444444
...